refactor(state): merge h5p into h5peditor
