+++ client/src/state/H5PEditor/H5PEditorActions.ts
@@ -1,4 +1,5 @@
-import * as H5PActions from '../h5p/H5PActions';
+import { ThunkAction, ThunkDispatch } from 'redux-thunk';
+
 import * as notifications from '../Notifications/NotificationsActions';
 
 import Logger from '../../helpers/Logger';
@@ -14,7 +15,31 @@ import {
     H5PEDITOR_LOADED,
     H5PEDITOR_SAVED,
     H5PPLAYER_INITIALIZED,
-    H5PEDITOR_SAVE_ERROR
+    H5PEDITOR_SAVE_ERROR,
+    DeleteActions,
+    LoadPlayerContentActions,
+    LoadEditorContentActions,
+    SaveContentActions,
+    ExportActions,
+    H5P_DELETE_ERROR,
+    H5P_DELETE_REQUEST,
+    H5P_DELETE_SUCCESS,
+    H5P_EXPORT_ERROR,
+    H5P_EXPORT_REQUEST,
+    H5P_EXPORT_SUCCESS,
+    H5P_IMPORT_ERROR,
+    H5P_IMPORT_REQUEST,
+    H5P_IMPORT_SUCCESS,
+    H5P_UPDATE_ERROR,
+    H5P_UPDATE_REQUEST,
+    H5P_UPDATE_SUCCESS,
+    H5P_LOADPLAYERCONTENT_REQUEST,
+    H5P_LOADPLAYERCONTENT_SUCCESS,
+    H5P_SAVECONTENT_REQUEST,
+    H5P_SAVECONTENT_SUCCESS,
+    H5P_SAVECONTENT_ERROR,
+    H5P_LOADEDITORCONTENT_REQUEST,
+    H5P_LOADEDITORCONTENT_SUCCESS
 } from './H5PEditorTypes';
 
 import superagent from 'superagent';
@@ -24,10 +49,11 @@ import upath from 'upath';
 import { track } from '../track/actions';
 import { IState } from '..';
 import * as selectors from './H5PEditorSelectors';
-import * as h5pActions from '../h5p/H5PActions';
 import shortid from 'shortid';
 
 import store from '../index';
+
+import * as api from './H5PApi';
 
 const log = new Logger('actions:tabs');
 
@@ -96,7 +122,7 @@ export function clickOnCloseTab(tabId: string): any {
         const contentId = selectors.tab(getState(), tabId).contentId;
 
         if (contentId) {
-            await dispatch(H5PActions.deleteH5P(contentId));
+            await dispatch(deleteH5P(contentId));
         }
 
         dispatch(closeTab(tabId));
@@ -124,7 +150,7 @@ export function clickOnFileInFiletree(name: string, path: string): any {
         // dispatch(openTab(tab));
 
         try {
-            const importAction = await dispatch(H5PActions.importH5P(path));
+            const importAction = await dispatch(importH5P(path));
 
             if (importAction.error) {
                 dispatch(
@@ -187,12 +213,7 @@ export function clickOnSaveButton(
         );
 
         const updateAction = await dispatch(
-            H5PActions.updateH5P(
-                params.params,
-                params.metadata,
-                library,
-                contentId
-            )
+            updateH5P(params.params, params.metadata, library, contentId)
         );
 
         if (updateAction.error) {
@@ -207,9 +228,7 @@ export function clickOnSaveButton(
 
         const updatedH5P = updateAction.payload;
 
-        const exportAction = await dispatch(
-            H5PActions.exportH5P(updatedH5P.id, path)
-        );
+        const exportAction = await dispatch(exportH5P(updatedH5P.id, path));
 
         if (exportAction.error) {
             if (exportAction.error.response.body.code === 'user-abort') {
@@ -270,12 +289,7 @@ export function updateH5PInTab(
         }
 
         const updateAction = await dispatch(
-            H5PActions.updateH5P(
-                params.params,
-                params.metadata,
-                library,
-                contentId
-            )
+            updateH5P(params.params, params.metadata, library, contentId)
         );
 
         if (updateAction.error) {
@@ -309,7 +323,7 @@ export function closeTab(id: string): any {
         const tab = selectors.tab(store.getState(), id);
 
         if (tab && tab.contentId) {
-            dispatch(h5pActions.deleteH5P(tab.contentId));
+            dispatch(deleteH5P(tab.contentId));
         }
 
         dispatch({
@@ -334,5 +348,246 @@ export function updateTab(
     return {
         payload: { tabId, update },
         type: H5PEDITOR_UPDATE_TAB
+    };
+}
+
+export function saveContent(
+    tabId: string,
+    contentId: string,
+    requestBody: { library: string; params: any }
+): ThunkAction<void, null, null, SaveContentActions> {
+    return async (dispatch: ThunkDispatch<null, null, SaveContentActions>) => {
+        dispatch({
+            payload: {
+                tabId,
+                library: requestBody.library,
+                params: requestBody.params
+            },
+            type: H5P_SAVECONTENT_REQUEST
+        });
+
+        try {
+            if (contentId) {
+                const response = JSON.parse(
+                    (await api.saveContent(contentId, requestBody)).text
+                );
+
+                dispatch({
+                    payload: {
+                        tabId,
+                        contentId: response.contentId,
+                        metadata: response.metadata
+                    },
+                    type: H5P_SAVECONTENT_SUCCESS
+                });
+
+                return response;
+            }
+
+            const response = JSON.parse(
+                (await api.createContent(requestBody)).text
+            );
+            dispatch({
+                payload: {
+                    tabId,
+                    contentId: response.contentId,
+                    metadata: response.metadata
+                },
+                type: H5P_SAVECONTENT_SUCCESS
+            });
+
+            return response;
+        } catch (error) {
+            dispatch({
+                payload: {
+                    tabId
+                },
+                type: H5P_SAVECONTENT_ERROR
+            });
+        }
+    };
+}
+
+export function loadEditorContent(
+    tabId: string,
+    contentId: ContentId
+): ThunkAction<void, null, null, LoadEditorContentActions> {
+    return async (
+        dispatch: ThunkDispatch<null, null, LoadEditorContentActions>
+    ) => {
+        dispatch({
+            payload: { contentId, tabId },
+            type: H5P_LOADEDITORCONTENT_REQUEST
+        });
+
+        try {
+            const content = JSON.parse(
+                (await api.loadEditorContent(contentId)).text
+            );
+
+            dispatch({
+                payload: { contentId, tabId, content },
+                type: H5P_LOADEDITORCONTENT_SUCCESS
+            });
+
+            return content;
+        } catch (error) {
+            throw new Error(error);
+        }
+    };
+}
+
+export function loadPlayerContent(
+    contentId: ContentId
+): ThunkAction<void, null, null, LoadPlayerContentActions> {
+    if (!contentId) {
+        throw new Error('no contentid');
+    }
+    return async (
+        dispatch: ThunkDispatch<null, null, LoadPlayerContentActions>
+    ) => {
+        dispatch({
+            payload: { contentId },
+            type: H5P_LOADPLAYERCONTENT_REQUEST
+        });
+
+        const content = JSON.parse(
+            (await api.loadPlayerContent(contentId)).text
+        );
+
+        dispatch({
+            payload: { contentId, content },
+            type: H5P_LOADPLAYERCONTENT_SUCCESS
+        });
+
+        return content;
+    };
+}
+
+export function deleteH5P(
+    contentId: ContentId
+): ThunkAction<void, null, null, DeleteActions> {
+    return (dispatch: ThunkDispatch<null, null, DeleteActions>) => {
+        dispatch({
+            payload: { contentId },
+            type: H5P_DELETE_REQUEST
+        });
+
+        return api
+            .deleteH5P(contentId)
+            .then((response) => {
+                return dispatch({
+                    payload: { contentId },
+                    type: H5P_DELETE_SUCCESS
+                });
+            })
+            .catch((error) => {
+                return dispatch({
+                    error,
+                    payload: { contentId },
+                    type: H5P_DELETE_ERROR
+                });
+            });
+    };
+}
+
+export function exportH5P(
+    contentId: ContentId,
+    path?: string
+): ThunkAction<void, null, null, ExportActions> {
+    return (dispatch: any) => {
+        dispatch({
+            payload: { id: contentId, path },
+            type: H5P_EXPORT_REQUEST
+        });
+
+        return api
+            .exportH5P(contentId, path)
+            .then(({ body }) =>
+                dispatch({
+                    // tslint:disable-next-line: object-shorthand-properties-first
+                    payload: { id: contentId, ...body },
+                    type: H5P_EXPORT_SUCCESS
+                })
+            )
+            .catch((error: Error) =>
+                dispatch({
+                    error,
+                    payload: { id: contentId, path },
+                    type: H5P_EXPORT_ERROR
+                })
+            );
+    };
+}
+
+export function importH5P(
+    path: string
+): ThunkAction<void, null, null, ExportActions> {
+    return (dispatch: any) => {
+        dispatch({
+            payload: { path },
+            type: H5P_IMPORT_REQUEST
+        });
+
+        return api
+            .importH5P(path)
+            .then(({ body }) =>
+                dispatch({
+                    payload: { path, h5p: body },
+                    type: H5P_IMPORT_SUCCESS
+                })
+            )
+            .catch((error: Error) =>
+                dispatch({
+                    error,
+                    payload: {
+                        path
+                    },
+                    type: H5P_IMPORT_ERROR
+                })
+            );
+    };
+}
+
+export function updateH5P(
+    parameters: any,
+    metadata: any,
+    library: string,
+    id?: ContentId
+): any {
+    return (dispatch: any) => {
+        dispatch({
+            payload: { parameters, metadata, library, id },
+            type: H5P_UPDATE_REQUEST
+        });
+
+        return api
+            .updateH5P(
+                {
+                    id,
+                    library,
+                    metadata,
+                    parameters
+                },
+                id
+            )
+            .then(({ body }) =>
+                dispatch({
+                    payload: { ...body },
+                    type: H5P_UPDATE_SUCCESS
+                })
+            )
+            .catch((error: Error) =>
+                dispatch({
+                    error,
+                    payload: {
+                        id,
+                        library,
+                        metadata,
+                        parameters
+                    },
+                    type: H5P_UPDATE_ERROR
+                })
+            );
     };
 }
